Updated order of tile selection from baf, moved strategy class to new file
